Move ABS, DomOverWDeg and impact search to the search strategies package
