+++ src/test/java/org/chocosolver/solver/search/ImpactTest.java
@@ -31,7 +31,7 @@ package org.chocosolver.solver.search;
 
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.Solver;
-import org.chocosolver.solver.search.strategy.selectors.variables.ImpactBased;
+import org.chocosolver.solver.search.strategy.strategy.ImpactBased;
 import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.util.ProblemMaker;
 import org.testng.annotations.Test;
